Update README.md
The bar has to be raised a bit: new contributed files that are of `sandbox`-quality should not come into the repository. We, as COSIMA community, ought to help the contributor to add some comments and explanations in the notebook so that it deserves to be places somewhere else other than the `sandbox`.