Fix embedded (non-plot) images in examples
We need to tell sphinx-nbexamples about the non-plot images in
notebooks, so that it can copy them to the right place. Because it
already uses the "images" directory for plots and thumbnails, this
seems to be the right place to have these images -- otherwise a copy
would fail due to a non-existent directory.

Closes #123.